polish

+++ base-reports/src/main/java/de/freese/base/reports/exporter/Exporter.java
@@ -25,14 +25,14 @@ public interface Exporter<T>
      *
      * @param progressCallback {@link ProgressCallback}, optional
      */
-    void export(OutputStream outputStream, BiConsumer<Long, Long> progressCallback, T model) throws Exception;
+    void export(OutputStream outputStream, BiConsumer<Integer, Integer> progressCallback, T model) throws Exception;
 
     /**
      * Erzeugt das Dokument und schreibt es in die Datei.
      *
      * @param progressCallback {@link ProgressCallback}, optional
      */
-    default void export(Path filePath, BiConsumer<Long, Long> progressCallback, T model) throws Exception
+    default void export(Path filePath, BiConsumer<Integer, Integer> progressCallback, T model) throws Exception
     {
         try (OutputStream outputStream = new BufferedOutputStream(Files.newOutputStream(filePath)))
         {
